Test out sqlite db for deployment (see body)
Heroku provides a Postgres plugin. The free edition has a limit on the amount of
concurrent connections. It is possible that at the scale we are currently using
running a SQLite DB may actually be more effective in the short term. However,
there are issues of migration and redeployment to consider.